web_app/backend/requirements.txt
fastapi
uvicorn
pyyaml
cohere
